src/requirements.txt
selenium==3.141.0
pandas==1.0.3
requests==2.23.0
PyHamcrest==2.0.2
allure-pytest==2.8.22
pytest==6.1.2
assertpy==1.1
pytest-html==3.0.0
